Fixes #716 - Modified exit checks. Converted boundary length to long.

+++ warp10/src/main/java/io/warp10/continuum/egress/HBaseStoreClient.java
@@ -160,7 +160,7 @@ public class HBaseStoreClient implements StoreClient {
   }
   
   @Override
-  public GTSDecoderIterator fetch(final ReadToken token, final List<Metadata> metadatas, final long now, final long then, long count, long skip, double sample, final boolean writeTimestamp, int preBoundary, int postBoundary) throws IOException {
+  public GTSDecoderIterator fetch(final ReadToken token, final List<Metadata> metadatas, final long now, final long then, long count, long skip, double sample, final boolean writeTimestamp, long preBoundary, long postBoundary) throws IOException {
 
     if (preBoundary < 0) {
       preBoundary = 0;
