+++ warp10/src/main/java/io/warp10/standalone/StandaloneParallelStoreClientWrapper.java
@@ -49,7 +49,7 @@ public class StandaloneParallelStoreClientWrapper implements StoreClient {
   }
   
   @Override
-  public GTSDecoderIterator fetch(ReadToken token, List<Metadata> metadatas, long now, long then, long count, long skip, double sample, boolean writeTimestamp, final int preBoundary, final int postBoundary) throws IOException {
+  public GTSDecoderIterator fetch(ReadToken token, List<Metadata> metadatas, long now, long then, long count, long skip, double sample, boolean writeTimestamp, final long preBoundary, final long postBoundary) throws IOException {
     if (ParallelGTSDecoderIteratorWrapper.useParallelScanners()) {
       return new ParallelGTSDecoderIteratorWrapper(parent, token, now, then, count, skip, sample, metadatas, preBoundary, postBoundary);
     } else {
